# specs/todo-app-spec/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo App API
  description: REST API for the Todo Full-Stack Web Application with JWT authentication
  version: 1.0.0
servers:
  - url: https://api.todoapp.com/v1
    description: Production server
  - url: https://staging-api.todoapp.com/v1
    description: Staging server

paths:
  /auth/register:
    post:
      summary: Register a new user
      description: Creates a new user account with the provided credentials
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      summary: Authenticate user
      description: Authenticates a user and returns a JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      summary: Logout user
      description: Invalidates the current user session
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Successfully logged out"

  /users/me:
    get:
      summary: Get current user
      description: Retrieves the currently authenticated user's information
      tags:
        - Users
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /tasks:
    get:
      summary: Get user's tasks
      description: Retrieves all tasks for the authenticated user with optional filtering
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter tasks by status
          required: false
          schema:
            type: string
            enum: [pending, in_progress, completed]
        - name: priority
          in: query
          description: Filter tasks by priority
          required: false
          schema:
            type: string
            enum: [low, medium, high]
        - name: search
          in: query
          description: Search tasks by title or description
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of tasks to return
          required: false
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          description: Number of tasks to skip for pagination
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of user's tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/TaskResponse'
                  total:
                    type: integer
                    description: Total number of tasks matching the criteria
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create a new task
      description: Creates a new task for the authenticated user
      tags:
        - Tasks
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /tasks/{id}:
    get:
      summary: Get a specific task
      description: Retrieves a specific task by its ID
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update a task
      description: Updates an existing task with the provided data
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete a task
      description: Deletes a specific task by its ID
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Task deleted successfully"
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserRegister:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          description: User's password (min 8 characters)
          example: securePassword123

    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        password:
          type: string
          format: password
          description: User's password
          example: securePassword123

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          description: User's unique identifier
          example: 123
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2023-10-20T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Account last update timestamp
          example: "2023-10-20T10:00:00Z"
        is_active:
          type: boolean
          description: Whether the account is active
          example: true

    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          description: Type of the token
          example: "bearer"
        user:
          $ref: '#/components/schemas/UserResponse'

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 200
          description: Task title
          example: "Complete project proposal"
        description:
          type: string
          maxLength: 1000
          description: Detailed task description
          example: "Write and review the Q4 project proposal document"
        priority:
          type: string
          enum: [low, medium, high]
          description: Task priority level
          example: "medium"
        due_date:
          type: string
          format: date-time
          description: Deadline for the task
          example: "2023-12-31T23:59:59Z"

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          description: Task title
          example: "Complete project proposal"
        description:
          type: string
          maxLength: 1000
          description: Detailed task description
          example: "Write and review the Q4 project proposal document"
        status:
          type: string
          enum: [pending, in_progress, completed]
          description: Task status
          example: "in_progress"
        priority:
          type: string
          enum: [low, medium, high]
          description: Task priority level
          example: "high"
        due_date:
          type: string
          format: date-time
          description: Deadline for the task
          example: "2023-12-31T23:59:59Z"

    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          description: Task's unique identifier
          example: 456
        user_id:
          type: integer
          description: ID of the user who owns the task
          example: 123
        title:
          type: string
          description: Task title
          example: "Complete project proposal"
        description:
          type: string
          description: Detailed task description
          example: "Write and review the Q4 project proposal document"
        status:
          type: string
          enum: [pending, in_progress, completed]
          description: Task status
          example: "in_progress"
        priority:
          type: string
          enum: [low, medium, high]
          description: Task priority level
          example: "high"
        due_date:
          type: string
          format: date-time
          description: Deadline for the task
          example: "2023-12-31T23:59:59Z"
        created_at:
          type: string
          format: date-time
          description: Task creation timestamp
          example: "2023-10-20T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Task last update timestamp
          example: "2023-10-21T14:30:00Z"
        completed_at:
          type: string
          format: date-time
          description: Timestamp when task was marked complete (null if not completed)
          example: null

    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Error message
          example: "Invalid credentials"
        error_code:
          type: string
          description: Specific error code
          example: "INVALID_CREDENTIALS"